+++ client/package.json
@@ -26,6 +26,7 @@
     "is-lite": "0.8.0",
     "js-cookie": "3.0.0",
     "nanoid": "3.1.23",
+    "node-sass": "4.14.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-headless-tabs": "3.0.1",
@@ -162,7 +163,7 @@
     "repo-tools": "0.2.2",
     "resolve": "1.20.0",
     "rsync": "0.6.1",
-    "sass-loader": "11.1.1",
+    "sass-loader": "10.1.1",
     "serve": "11.3.2",
     "size-limit": "4.10.2",
     "source-map-support": "0.5.19",
