+++ client/src/components/index.ts
@@ -6,4 +6,4 @@ export { default as Loader } from './Loader/Loader.component';
 export { default as Page } from './Page/Page.component';
 export { default as Paper } from './Paper/Paper.component';
 export { default as Separator } from './Separator/Separator.component';
-export { default as Table } from './Table/Table.component';
+export * from './Table/Table.component';
