+++ client/src/containers/Settings/ReposSettings/AllRepos/AllRepos.view.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { PlusCircleIcon } from 'icons';
+import cl from './AllRepos.module.scss';
+
+export const RepoAddBtn = () => {
+  return (
+    <button className={cl.repoAddBtn}>
+      <PlusCircleIcon className={cl.repoAddBtnIcon} />
+    </button>
+  );
+};
