+++ client/src/containers/Settings/ReposSettings/ReposSettings.module.css
@@ -1,55 +1,3 @@
 .main {
   @apply h-full flex space-x-4 overflow-hidden;
 }
-
-.repos {
-  @apply w-3/4 flex flex-col space-y-2;
-}
-
-.repos-table {
-  @apply h-full flex overflow-hidden;
-}
-
-.repos-actions {
-  @apply flex space-x-2 justify-end;
-}
-
-.user-repos {
-  @apply w-1/4 flex flex-col space-y-2;
-}
-
-.user-repos-table {
-  @apply h-full flex overflow-hidden;
-}
-
-.user-repos-actions {
-  @apply flex space-x-2 justify-end;
-}
-
-.repo-add-btn {
-  @apply h-6 w-6 focus:outline-none float-right;
-}
-
-.repo-remove-btn {
-  @apply h-6 w-6 focus:outline-none;
-}
-
-.repo-add-icon {
-  @apply h-6 w-6 text-green-700;
-}
-
-.repo-remove-icon {
-  @apply h-6 w-6 text-red-700;
-}
-
-:global(.dark) .repo-add-icon {
-  @apply text-green-500;
-}
-
-:global(.dark) .repo-remove-icon {
-  @apply text-red-500;
-}
-
-.user-repo-name {
-  @apply flex space-x-2 items-center;
-}
